PM Registration : grade benefit ticker and terms

+++ features/merchant/power_merchant_subscribe/src/main/java/com/tokopedia/power_merchant/subscribe/view/viewcomponent/PMProTargetView.kt
@@ -39,7 +39,7 @@ class PMProTargetView : ConstraintLayout {
     }
 
     fun showInfo(
-        completedOrder: Int,
+        completedOrder: Long,
         netIncome: Long,
         shopLevel: Int
     ) {
